Fix passing `--debug`(`--release`) when there's no such build config
At the moment, when `--debug` option is passed and there's no `.debug.abproject` in the project, CLI fails with error:
```
Operation cannot be completed because configurations Debug are invalid.
```

In the previous versions of CLI, we have been creating the file manually and the operation continues.
The new behavior is introduced with a change in `mobile-cli-lib` where it has been causing other issues.
Return the old behavior only for AppBuilder CLI by placing the deleted code in its source. Add check if `projectDir` is set, so the creation of project (which also uses the same code) will not fail.