Add message for discontinuation of Telerik Platform
Add message and change the code in CLI's entry point. From now on, each executed command will print message for the discontinuation and will not execute any other action.
Keep the logic for checking if the current version is the latest available, so we can release a new version of required.